feat(cli): nodejs loader (#515)
<!--
Pull requests are squash merged using:
- their title as the commit message
- their description as the commit body

Having a good title and description is important for the users to get
readable changelog and understand when they need to update his code and
how.
-->

### Describe your change

Add support for using nodejs runtime to execute and serialize typescript
based typegraphs. This also adds support for `MCLI_LOADER_CMD` that can
be used to override the command to exec the typegraphs.

### Motivation and context

Previously, `meta-cli` either used the and `python` & `deno` runtimes to
serialize the typegraphs. Now that `@typegraph/sdk` also supports
Node.js, users might be developing in environments wher `deno` runtime
is not availaible but `node` is. This PR provides a way fwd in those
cases.

### Migration notes

<!-- Explain HOW users should update their code when required -->

### Checklist

- [ ] The change come with new or modified tests
- [x] Hard-to-understand functions have explanatory comments
- [ ] End-user documentation is updated to reflect the change