feat: Customizable oauth2 profiler (#538)
<!--
Pull requests are squash merged using:
- their title as the commit message
- their description as the commit body

Having a good title and description is important for the users to get
readable changelog and understand when they need to update his code and
how.
-->

### Describe your change

Enable custom profiler for the std (predefined) Oauth2 providers:
- Default profiler
- No profiler
- Extended default profiler
- Custom profiler

### Motivation and context

We may want for example to add the Github login in the profile in
addition to the id.

### Migration notes

_No migration needed._

### Checklist

- [ ] The change come with new or modified tests
- [ ] Hard-to-understand functions have explanatory comments
- [ ] End-user documentation is updated to reflect the change