fix(typegate): fix `tmp` dir issue, bump to 0.3.4 (#583)
- Fixes an issue with the prisma migration code being unable to properly
`mktmpd` when running in the `typegate` images.
- Removes `cross` usage for all but the `aarch64-unknown-linux-gnu`
target in the `meta-cli` release jobs. Deno doesn't like being cross
compiled, doesn't like being built by the cross toolchain in general
(even for the native target). The consequence of all this is that the
`aarch64-unknown-linux-gnu` target will temporarily lack the `full`
version builds.
- This also bumps the version of the 0.3.4 to get the fix out.

#### Motivation and context

The main `TMP_DIR` was not created properly in the `typegate`
`Dockerfile`.

#### Migration notes

_No changes required_

### Checklist

- [ ] The change come with new or modified tests
- [ ] Hard-to-understand functions have explanatory comments
- [ ] End-user documentation is updated to reflect the change