feat(subs,gate): port agent concept (#845)
Continuation of #844 

Note: 
There is currently no need for additional flags when running typegate,
existing runs should be picked automatically as long as the typegraph
owning the workflow is deployed. SubstantialRuntime will launch and
start the agent, the later will attempt to 'pick' jobs on the backend
and spawn workers if it can acquire it.
Workflows are identified relative to their names for now, the content
hash might be the way to go going forward, downside being that it will
change even on smaller modifications (e.g. spaces, missing semicolons,
..).

#### Migration notes

None

- [x] The change comes with new or modified tests
- [ ] Hard-to-understand functions have explanatory comments
- [ ] End-user documentation is updated to reflect the change